+++ src/components/BoxMotion/index.ts
@@ -1,0 +1,2 @@
+export * from "./BoxMotion";
+export * from "./BoxMotion.types";